add keep flag for paperman diff

+++ paperman/subcommands/diff.py
@@ -156,11 +156,15 @@ def main(args):
 
   if os.path.exists(target):
     shutil.move(os.path.join(buildDir, 'paperman-diff.pdf'), args.outfile)
-    for f in os.listdir(buildDir):
-      if f.startswith('paperman-diff'):
-        fname = os.path.join(buildDir, f)
-        os.remove(fname)
-        io.verb(f'removed {fname}')
+
+    # remove build files if 'keep' flag is not set
+    if not args.keep:
+      for f in os.listdir(buildDir):
+        if f.startswith('paperman-diff'):
+          fname = os.path.join(buildDir, f)
+          os.remove(fname)
+          io.verb(f'removed {fname}')
+
     io.info(f'successfully generated {args.outfile}')
   else:
     io.err('failed to build diff')
